app/src/main/res/values-pt-rBR
This String was very outdated, so I took the liberty to update it. Since this is now my primary application for reading manga. maybe there are some adjustments to be made in the translation, but I won't be able to do it without a compiled version in the application to test.

Translation based on the String of this link:
https://github.com/Jays2Kings/tachiyomiJ2K/blob/master/app/src/main/res/values/strings.xml